Rename: Skeleton폴더 내 파일 경로 변경 (#12)
