+++ src/Components/common/Skeleton/HomePostSkeleton.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import SkeletonItem from '../../Styles/SkeletonItem';
-import UserSkeleton from '../common/UserSkeleton';
+import SkeletonItem from '../../../Styles/SkeletonItem';
+import UserSkeleton from './common/UserSkeleton';
 
 const HomePostSkeleton = () => {
   return (
